material_colors::dynamic_color::contrast_curve

Struct ContrastCurve

pub struct ContrastCurve {
    pub low: f64,
    pub normal: f64,
    pub medium: f64,
    pub high: f64,
}

A class containing a value that changes with the contrast level.

Usually represents the contrast requirements for a dynamic color on its background. The four values correspond to values for contrast levels -1.0, 0.0, 0.5, and 1.0, respectively.

Fields
low: f64normal: f64medium: f64high: f64
Implementations

impl ContrastCurve

pub fn get(&self, contrast_level: f64) -> f64

Returns the value at a given contrast level.

Parameter contrastLevel: The contrast level. 0.0 is the default (normal); -1.0 is the lowest; 1.0 is the highest.

Returns: The value. For contrast ratios, a number between 1.0 and 21.0.