material_colors::dynamic_color

Module dynamic_scheme

Macros

rgb_getters 🔒

Structs

DynamicScheme

Constructed by a set of values representing the current UI state and provides a set of TonalPalettes that can create colors that fit in with the theme style.

Enums

Platform

The platform on which a scheme is intended to be used.

SchemePalette

Identifies one of the six palettes owned by a DynamicScheme.

Constants

DEFAULT_PLATFORM

DEFAULT_SPEC_VERSION

Functions

sanitize_degrees 🔒