Struct material_colors::dynamic_color::DynamicColor

pub struct DynamicColor {
    pub name: String,
    palette: Box<fn(_: &DynamicScheme) -> &TonalPalette>,
    tone: Box<fn(_: &DynamicScheme) -> f64>,
    is_background: bool,
    background: Option<Box<fn(_: &DynamicScheme) -> DynamicColor>>,
    second_background: Option<Box<fn(_: &DynamicScheme) -> DynamicColor>>,
    contrast_curve: Option<ContrastCurve>,
    tone_delta_pair: Option<Box<fn(_: &DynamicScheme) -> ToneDeltaPair>>,
    _hct_cache: HashMap<DynamicScheme, Hct>,
}

A color that adjusts itself based on UI state provided by DynamicScheme.

This color automatically adjusts to accommodate a desired contrast level, or other adjustments such as differing in light mode versus dark mode, or what the theme is, or what the color that produced the theme is, etc.

Colors without backgrounds do not change tone when contrast changes. Colors with backgrounds become closer to their background as contrast lowers, and further when contrast increases.

Prefer the static constructors. They provide a much more simple interface, such as requiring just a hexcode, or just a hexcode and a background.

Ultimately, each component necessary for calculating a color, adjusting it for a desired contrast level, and ensuring it has a certain lightness/tone difference from another color, is provided by a function that takes a DynamicScheme and returns a value. This ensures ultimate flexibility, any desired behavior of a color for any design system, but it usually unnecessary. See the default constructor for more information.

Fields
name: Stringpalette: Box<fn(_: &DynamicScheme) -> &TonalPalette>tone: Box<fn(_: &DynamicScheme) -> f64>is_background: boolbackground: Option<Box<fn(_: &DynamicScheme) -> DynamicColor>>second_background: Option<Box<fn(_: &DynamicScheme) -> DynamicColor>>contrast_curve: Option<ContrastCurve>tone_delta_pair: Option<Box<fn(_: &DynamicScheme) -> ToneDeltaPair>>_hct_cache: HashMap<DynamicScheme, Hct>
Implementations

impl DynamicColor

pub fn new<T: Into<String>>( name: T, palette: fn(_: &DynamicScheme) -> &TonalPalette, tone: fn(_: &DynamicScheme) -> f64, is_background: bool, background: Option<fn(_: &DynamicScheme) -> Self>, second_background: Option<fn(_: &DynamicScheme) -> Self>, contrast_curve: Option<ContrastCurve>, tone_delta_pair: Option<fn(_: &DynamicScheme) -> ToneDeltaPair>, ) -> Self

The base constructor for DynamicColor.

Strongly prefer using one of the convenience constructors. This class is arguably too flexible to ensure it can support any scenario. Functional arguments allow overriding without risks that come with subclasses.

For example, the default behavior of adjust tone at max contrast to be at a 7.0 ratio with its background is principled and matches accessibility guidance. That does not mean it’s the desired approach for every design system, and every color pairing, always, in every case.

Parameters:

name: The name of the dynamic color.

palette: Function that provides a TonalPalette given DynamicScheme. A TonalPalette is defined by a hue and chroma, so this replaces the need to specify hue/chroma. By providing a tonal palette, when contrast adjustments are made, intended chroma can be preserved.

tone: Function that provides a tone, given a DynamicScheme.

isBackground: Whether this dynamic color is a background, with some other color as the foreground.

background: The background of the dynamic color (as a function of a DynamicScheme), if it exists.

secondBackground: A second background of the dynamic color (as a function of a DynamicScheme), if it exists.

contrastCurve: A ContrastCurve object specifying how its contrast against its background should behave in various contrast levels options.

toneDeltaPair: A ToneDeltaPair object specifying a tone delta constraint between two colors. One of them must be the color being constructed.

Unlikely to be useful unless a design system has some distortions where colors that don’t have a background/foreground relationship don’t want to have a formal relationship or a principled value for their tone distance based on common contrast / tone delta values, yet, want tone distance.

pub fn get_argb(&mut self, scheme: &DynamicScheme) -> Argb

Return a Argb integer (i.e. a hex code).

Parameter scheme: Defines the conditions of the user interface, for example, whether or not it is dark mode or light mode, and what the desired contrast level is.

Returns: The color as an integer (Argb).

pub fn get_hct(&mut self, scheme: &DynamicScheme) -> Hct

Parameter scheme: Defines the conditions of the user interface, for example, whether or not it is dark mode or light mode, and what the desired contrast level is.

Returns: a color, expressed in the HCT color space, that this DynamicColor is under the conditions in scheme.

pub fn get_tone(&self, scheme: &DynamicScheme) -> f64

Parameter scheme: Defines the conditions of the user interface, for example, whether or not it is dark mode or light mode, and what the desired contrast level is.

Returns: a tone, T in the HCT color space, that this DynamicColor is under the conditions in scheme.

pub fn foreground_tone(bg_tone: f64, ratio: f64) -> f64

Given a background tone, find a foreground tone, while ensuring they reach a contrast ratio that is as close to ratio as possible.

Parameters:

bgTone: Tone in HCT. Range is 0 to 100, undefined behavior when it falls outside that range.

ratio: The contrast ratio desired between bgTone and the return value.

Returns: The desired foreground tone.

pub fn enable_light_foreground(tone: f64) -> f64

Adjusts a tone such that white has 4.5 contrast, if the tone is reasonably close to supporting it.

Parameter tone: The tone to be adjusted.

Returns: The (possibly updated) tone.

pub fn tone_prefers_light_foreground(tone: f64) -> bool

Returns whether tone prefers a light foreground.

People prefer white foregrounds on ~T60-70. Observed over time, and also by Andrew Somers during research for APCA.

T60 used as to create the smallest discontinuity possible when skipping down to T49 in order to ensure light foregrounds.

Since tertiaryContainer in dark monochrome scheme requires a tone of 60, it should not be adjusted. Therefore, 60 is excluded here.

Parameter tone: The tone to be judged.

Returns: whether tone prefers a light foreground.

pub fn tone_allows_light_foreground(tone: f64) -> bool

Returns whether tone can reach a contrast ratio of 4.5 with a lighter color.

Parameter tone: The tone to be judged.

Returns: whether tone allows a light foreground.