material_colors::hct::cam16

Struct Cam16

pub struct Cam16 {
    pub hue: f64,
    pub chroma: f64,
    pub j: f64,
    pub q: f64,
    pub m: f64,
    pub s: f64,
    pub jstar: f64,
    pub astar: f64,
    pub bstar: f64,
}

CAM16, a color appearance model. Colors are not just defined by their hex code, but rather, a hex code and viewing conditions.

CAM16 instances also have coordinates in the CAM16-UCS space, called J*, a*, b*, or jstar, astar, bstar in code. CAM16-UCS is included in the CAM16 specification, and should be used when measuring distances between colors.

In traditional color spaces, a color can be identified solely by the observer’s measurement of the color. Color appearance models such as CAM16 also use information about the environment where the color was observed, known as the viewing conditions.

For example, white under the traditional assumption of a midday sun white point is accurately measured as a slightly chromatic blue by (roughly, hue 203, chroma 3, lightness 100) CAM16, a color appearance model. Colors are not just defined by their hex code, but rather, a hex code and viewing conditions.

CAM16 instances also have coordinates in the CAM16-UCS space, called J*, a*, b*, or jstar, astar, bstar in code. CAM16-UCS is included in the CAM16 specification, and should be used when measuring distances between colors.

In traditional color spaces, a color can be identified solely by the observer’s measurement of the color. Color appearance models such as CAM16 also use information about the environment where the color was observed, known as the viewing conditions.

For example, white under the traditional assumption of a midday sun white point is accurately measured as a slightly chromatic blue by (roughly, hue 203, chroma 3, lightness 100)

Fields
hue: f64

Like red, orange, yellow, green, etc.

chroma: f64

Informally, colorfulness / color intensity. Like saturation in HSL, except perceptually accurate.

j: f64

Lightness

q: f64

Brightness; ratio of lightness to white point’s lightness

m: f64

Colorfulness

s: f64

Saturation; ratio of chroma to white point’s chroma

jstar: f64

CAM16-UCS J coordinate

astar: f64

CAM16-UCS a coordinate

bstar: f64

CAM16-UCS b coordinate

Implementations

impl Cam16

pub fn distance(&self, other: &Self) -> f64

CAM16 instances also have coordinates in the CAM16-UCS space, called J*, a*, b*, or jstar, astar, bstar in code. CAM16-UCS is included in the CAM16 specification, and should be used when measuring distances between colors.

pub fn from_rgb_in_viewing_conditions( rgb: Rgb, viewing_conditions: &ViewingConditions, ) -> Self

Given color expressed in Rgb form and viewed in viewing_conditions, convert to Cam16.

Panics

Will panic if the hue is between 0 and 360

pub fn from_xyz_in_viewing_conditions( value: Xyz, viewing_conditions: &ViewingConditions, ) -> Self

Given color expressed in Xyz form and viewed in viewing_conditions, convert to Cam16.

Panics

Will panic if the hue is between 0 and 360

pub fn from_jch(j: f64, c: f64, h: f64) -> Self

Create a CAM16 color from lightness j, chroma c, and hue h, assuming the color was viewed in default viewing conditions.

pub fn from_jch_in_viewing_conditions( j: f64, c: f64, h: f64, viewing_conditions: &ViewingConditions, ) -> Self

Create a CAM16 color from lightness j, chroma c, and hue h, in viewing_conditions.

pub fn from_ucs(jstar: f64, astar: f64, bstar: f64) -> Self

Create a CAM16 color from CAM16-UCS coordinates jstar, astar, bstar. assuming the color was viewed in default viewing conditions.

pub fn from_ucs_in_viewing_conditions( jstar: f64, astar: f64, bstar: f64, viewing_conditions: &ViewingConditions, ) -> Self

Create a CAM16 color from CAM16-UCS coordinates jstar, astar, bstar. in viewing_conditions.

pub fn viewed(&self, viewing_conditions: &ViewingConditions) -> Rgb

Rgb representation of a color, given the color was viewed in viewing_conditions

pub fn xyz_in_viewing_conditions( &self, viewing_conditions: &ViewingConditions, ) -> Xyz

Xyz representation of CAM16 seen in viewing_conditions.