material_colors::hct::solver

Struct HctSolver

pub struct HctSolver;

Implementations

impl HctSolver

const fn sanitize_radians(angle: f64) -> f64

Sanitizes a small enough angle in radians.

angle: An angle in radians; must not deviate too much from 0.

Returns A coterminal angle between 0 and 2pi.

fn true_delinearized(rgb_component: f64) -> f64

Delinearizes an Rgb component, returning a floating-point number.

0.0 <= rgb_component <= 100.0 represents linear R/G/B channel. 0.0 <= output <= 255.0, color channel converted to regular Rgb space.

fn chromatic_adaptation(component: f64) -> f64

fn hue_of(linrgb: [f64; 3]) -> f64

Returns the hue of linrgb, a linear Rgb color, in CAM16, in radians.

fn are_in_cyclic_order(a: f64, b: f64, c: f64) -> bool

fn intercept(source: f64, mid: f64, target: f64) -> f64

Solves the lerp equation.

Returns a number t such that lerp(source, target, t) = mid.

fn lerp_point(source: [f64; 3], t: f64, target: [f64; 3]) -> [f64; 3]

fn set_coordinate( source: [f64; 3], coordinate: f64, target: [f64; 3], axis: usize, ) -> [f64; 3]

i32ersects a segment with a plane.

Returns the intersection point of:

the segment with source and target as its endpoints, and

the plane … R = coordinate if axis == 0 … G = coordinate if axis == 1 … B = coordinate if axis == 2

fn is_bounded(x: f64) -> bool

fn nth_vertex(y: f64, n: i32) -> [f64; 3]

Returns the nth possible vertex of the polygonal intersection.

Given a plane Y = y and an zero-based index n such that 0 <= n <= 11, returns the nth possible vertex of the polygonal intersection of the plane and the Rgb cube, in linear Rgb coordinates, if it exists. If this possible vertex lies outside of the cube, [-1.0, -1.0, -1.0] is returned.

fn bisect_to_segment(y: f64, target_hue: f64) -> [[f64; 3]; 2]

Finds the segment containing the desired color.

Given a plane Y = y and a desired target_hue, returns the segment containing the desired color, represented as an array of its two endpoints.

const fn midpoint(a: [f64; 3], b: [f64; 3]) -> [f64; 3]

const fn critical_plane_below(x: f64) -> i16

const fn critical_plane_above(x: f64) -> i16

fn bisect_to_limit(y: f64, target_hue: f64) -> [f64; 3]

Finds a color with the given Y and hue on the boundary of the cube.

Returns the color with the desired Y value y and hue target_hue, in linear Rgb coordinates.

fn inverse_chromatic_adaptation(adapted: f64) -> f64

fn find_result_by_j(hue_radians: f64, chroma: f64, y: f64) -> Rgb

Finds a color with the given hue, chroma, and Y.

Returns a color with the desired hue_radians, chroma, and y as a hexadecimal integer, if found; and returns 0 otherwise.

pub fn solve_to_rgb(hue_degrees: f64, chroma: f64, lstar: f64) -> Rgb

Finds an sRgb color with the given hue, chroma, and L*, if possible.

Returns a hexadecimal representing a sRgb color with its hue, chroma, and L* sufficiently close to hue_degrees, chroma, and lstar, respectively. If it is impossible to satisfy all three constraints, the hue and L* will be sufficiently close, and the chroma will be maximized.

pub fn solve_to_cam(hue_degrees: f64, chroma: f64, lstar: f64) -> Cam16

Finds a CAM16 object with the given hue, chroma, and L*, if possible.

Returns a CAM16 object representing a sRgb color with its hue, chroma, and L* sufficiently close to hue_degrees, chroma, and lstar, respectively. If it is impossible to satisfy all three constraints, the hue and L* will be sufficiently close, and the chroma will be maximized.