Struct material_colors::hct::viewing_conditions::ViewingConditions

pub struct ViewingConditions {

    pub white_point: [f64; 3],
    pub adapting_luminance: f64,
    pub background_lstar: f64,
    pub surround: f64,
    pub discounting_illuminant: bool,
    pub background_ytowhite_point_y: f64,
    pub aw: f64,
    pub nbb: f64,
    pub ncb: f64,
    pub c: f64,
    pub n_c: f64,
    pub drgb_inverse: [f64; 3],
    pub rgb_d: [f64; 3],
    pub fl: f64,
    pub f_lroot: f64,
    pub z: f64,

}

In traditional color spaces, a color can be identified solely by the observer’s measurement of the color. Color appearance models such as CAM16 also use information about the environment where the color was observed, known as the viewing conditions.

For example, white under the traditional assumption of a midday sun white point is accurately measured as a slightly chromatic blue by CAM16. (roughly, hue 203, chroma 3, lightness 100)

This class caches intermediate values of the CAM16 conversion process that depend only on viewing conditions, enabling speed ups.

Fields
white_point: [f64; 3]adapting_luminance: f64background_lstar: f64surround: f64discounting_illuminant: boolbackground_ytowhite_point_y: f64aw: f64nbb: f64ncb: f64c: f64n_c: f64drgb_inverse: [f64; 3]rgb_d: [f64; 3]fl: f64f_lroot: f64z: f64
Implementations

impl ViewingConditions

pub fn standard() -> Self

pub fn s_rgb() -> Self

pub fn make( white_point: Option<[f64; 3]>, adapting_luminance: Option<f64>, background_lstar: Option<f64>, surround: Option<f64>, discounting_illuminant: Option<bool>, ) -> Self

Convenience constructor for ViewingConditions.

Parameters affecting color appearance include:

whitePoint - coordinates of white in Xyz color space.

adaptingLuminance - light strength, in lux.

backgroundLstar - average luminance of 10 degrees around color.

surround - brightness of the entire environment.

discountingIlluminant - whether eyes have adjusted to lighting.

Panics

Will panic if the surround is equal or between 0.0 and 2.0