material_colors::palette

Module core

Structs

CorePaletteDeprecated

An intermediate concept between the key color for a UI theme, and a full color scheme. 5 tonal palettes are generated, all except one use the same hue as the key color, and all vary in chroma.

CorePalettes

Comprises foundational palettes to build a color scheme. Generated from a source color, these palettes will then be part of a DynamicScheme together with appearance preferences.