Module material_colors::scheme::variant

Modules

content 🔒

expressive 🔒

fidelity 🔒

fruit_salad 🔒

monochrome 🔒

neutral 🔒

rainbow 🔒

tonal_spot 🔒

vibrant 🔒

Structs

SchemeContent

SchemeExpressive

A Dynamic Color theme that is intentionally detached from the input color.

SchemeFidelity

SchemeFruitSalad

SchemeMonochrome

SchemeNeutral

SchemeRainbow

SchemeTonalSpot

SchemeVibrant

A Dynamic Color theme that is intentionally detached from the input color.