material_colors::theme

Struct ThemeBuilder

pub struct ThemeBuilder {
    source: Argb,
    variant: Variant,
    color_match: bool,
    primary: Option<Argb>,
    secondary: Option<Argb>,
    tertiary: Option<Argb>,
    error: Option<Argb>,
    neutral: Option<Argb>,
    neutral_variant: Option<Argb>,
    custom_colors: Vec<CustomColor>,
}

Fields
source: Argbvariant: Variantcolor_match: boolprimary: Option<Argb>secondary: Option<Argb>tertiary: Option<Argb>error: Option<Argb>neutral: Option<Argb>neutral_variant: Option<Argb>custom_colors: Vec<CustomColor>
Implementations

impl ThemeBuilder

pub const fn with_source(source: Argb) -> Self

Creates a theme builder with a custom source color.

pub const fn variant(self, variant: Variant) -> Self

Sets the theme variant.

pub const fn primary(self, color: Argb) -> Self

Sets the primary color of the theme.

pub const fn secondary(self, color: Argb) -> Self

Sets the secondary color of the theme.

pub const fn tertiary(self, color: Argb) -> Self

Sets the tertiary color of the theme.

pub const fn error(self, color: Argb) -> Self

Sets the error color of the theme.

pub const fn neutral(self, color: Argb) -> Self

Sets the neutral color, used for background and surfaces.

pub const fn neutral_variant(self, color: Argb) -> Self

Sets the neutral variant color, used for for medium emphasis and variants.

pub fn custom_colors(self, custom_colors: Vec<CustomColor>) -> Self

Sets the custom colors, used as complementary tones.

Custom colors are also known as extended colors.

pub const fn color_match(self, enabled: bool) -> Self

pub fn build(self) -> Theme