material_colors::dynamic_color

Struct DynamicColor

pub struct DynamicColor {
    extended_data: Option<ExtendedColorData>,
    pub name: &'static str,
    palette_: for<'a> fn(Option<ExtendedColorData>, &'a DynamicScheme) -> &'a TonalPalette,
    tone_: fn(Option<ExtendedColorData>, &DynamicScheme) -> f64,
    is_background: bool,
    chroma_multiplier_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<f64>,
    background_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<&'static Self>,
    second_background_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<&'static Self>,
    contrast_curve_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<ContrastCurve>,
    tone_delta_pair_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<ToneDeltaPair>,
    opacity_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<f64>,
}

A color that adjusts itself based on UI state provided by DynamicScheme.

This color automatically adjusts to accommodate a desired contrast level, or other adjustments such as differing in light mode versus dark mode, or what the theme is, or what the color that produced the theme is, etc.

Colors without backgrounds do not change tone when contrast changes. Colors with backgrounds become closer to their background as contrast lowers, and further when contrast increases.

Prefer the static constructors. They provide a much more simple interface, such as requiring just a hexcode, or just a hexcode and a background.

Ultimately, each component necessary for calculating a color, adjusting it for a desired contrast level, and ensuring it has a certain lightness/tone difference from another color, is provided by a function that takes a DynamicScheme and returns a value. This ensures ultimate flexibility, any desired behavior of a color for any design system, but it usually unnecessary. See the default constructor for more information.

Fields
extended_data: Option<ExtendedColorData>name: &'static strpalette_: for<'a> fn(Option<ExtendedColorData>, &'a DynamicScheme) -> &'a TonalPalettetone_: fn(Option<ExtendedColorData>, &DynamicScheme) -> f64is_background: boolchroma_multiplier_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<f64>background_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<&'static Self>second_background_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<&'static Self>contrast_curve_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<ContrastCurve>tone_delta_pair_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<ToneDeltaPair>opacity_: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<f64>
Implementations

impl DynamicColor

pub const fn foreground_color( name: &'static str, palette: for<'a> fn(Option<ExtendedColorData>, &'a DynamicScheme) -> &'a TonalPalette, tone: fn(Option<ExtendedColorData>, &DynamicScheme) -> f64, ) -> Self

The base constructor for DynamicColor.

Strongly prefer using one of the convenience constructors. This class is arguably too flexible to ensure it can support any scenario. Functional arguments allow overriding without risks that come with subclasses.

For example, the default behavior of adjust tone at max contrast to be at a 7.0 ratio with its background is principled and matches accessibility guidance. That does not mean it’s the desired approach for every design system, and every color pairing, always, in every case.

Parameters:

name: The name of the dynamic color.

palette: Function that provides a TonalPalette given DynamicScheme. A TonalPalette is defined by a hue and chroma, so this replaces the need to specify hue/chroma. By providing a tonal palette, when contrast adjustments are made, intended chroma can be preserved.

tone: Function that provides a tone, given a DynamicScheme.

isBackground: Whether this dynamic color is a background, with some other color as the foreground.

background: The background of the dynamic color (as a function of a DynamicScheme), if it exists.

secondBackground: A second background of the dynamic color (as a function of a DynamicScheme), if it exists.

contrastCurve: A ContrastCurve object specifying how its contrast against its background should behave in various contrast levels options.

toneDeltaPair: A ToneDeltaPair object specifying a tone delta constraint between two colors. One of them must be the color being constructed.

Unlikely to be useful unless a design system has some distortions where colors that don’t have a background/foreground relationship don’t want to have a formal relationship or a principled value for their tone distance based on common contrast / tone delta values, yet, want tone distance.

pub const fn background_color( name: &'static str, palette: for<'a> fn(Option<ExtendedColorData>, &'a DynamicScheme) -> &'a TonalPalette, tone: fn(Option<ExtendedColorData>, &DynamicScheme) -> f64, ) -> Self

pub fn tone(&self, scheme: &DynamicScheme) -> f64

pub fn palette<'a>(&self, scheme: &'a DynamicScheme) -> &'a TonalPalette

pub fn background(&self, scheme: &DynamicScheme) -> Option<&'static Self>

pub fn second_background(&self, scheme: &DynamicScheme) -> Option<&'static Self>

pub fn contrast_curve(&self, scheme: &DynamicScheme) -> Option<ContrastCurve>

pub fn tone_delta_pair(&self, scheme: &DynamicScheme) -> Option<ToneDeltaPair>

pub fn chroma_multiplier(&self, scheme: &DynamicScheme) -> Option<f64>

pub fn opacity(&self, scheme: &DynamicScheme) -> Option<f64>

pub const fn with_name(self, name: &'static str) -> Self

pub const fn with_tone( self, func: fn(Option<ExtendedColorData>, &DynamicScheme) -> f64, ) -> Self

pub const fn with_palette( self, func: for<'a> fn(Option<ExtendedColorData>, &'a DynamicScheme) -> &'a TonalPalette, ) -> Self

pub const fn with_background( self, func: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<&'static Self>, ) -> Self

pub const fn with_second_background( self, func: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<&'static Self>, ) -> Self

pub const fn with_contrast_curve( self, curve: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<ContrastCurve>, ) -> Self

pub const fn with_tone_delta_pair( self, func: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<ToneDeltaPair>, ) -> Self

pub const fn with_chroma_multiplier( self, func: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<f64>, ) -> Self

pub const fn with_opacity( self, func: fn(Option<ExtendedColorData>, &DynamicScheme) -> Option<f64>, ) -> Self

pub fn foreground_tone(bg_tone: f64, ratio: f64) -> f64

Given a background tone, find a foreground tone, while ensuring they reach a contrast ratio that is as close to ratio as possible.

Parameters:

bgTone: Tone in HCT. Range is 0 to 100, undefined behavior when it falls outside that range.

ratio: The contrast ratio desired between bgTone and the return value.

Returns: The desired foreground tone.

pub fn get_rgb(&self, scheme: &DynamicScheme) -> Rgb

pub fn get_hct(&self, scheme: &DynamicScheme) -> Hct

pub fn get_tone(&self, scheme: &DynamicScheme) -> f64

const fn validate_extended_color( &self, _: SpecVersion, extended_color: &'static Self, )

pub const fn extend_spec_version( &'static self, spec_version: SpecVersion, extended_color: &'static Self, ) -> Self

pub fn enable_light_foreground(tone: f64) -> f64

Adjusts a tone such that white has 4.5 contrast, if the tone is reasonably close to supporting it.

Parameter tone: The tone to be adjusted.

Returns: The (possibly updated) tone.

pub fn tone_prefers_light_foreground(tone: f64) -> bool

Returns whether tone prefers a light foreground.

People prefer white foregrounds on ~T60-70. Observed over time, and also by Andrew Somers during research for APCA.

T60 used as to create the smallest discontinuity possible when skipping down to T49 in order to ensure light foregrounds.

Since tertiaryContainer in dark monochrome scheme requires a tone of 60, it should not be adjusted. Therefore, 60 is excluded here.

Parameter tone: The tone to be judged.

Returns: whether tone prefers a light foreground.

pub fn tone_allows_light_foreground(tone: f64) -> bool

Returns whether tone can reach a contrast ratio of 4.5 with a lighter color.

Parameter tone: The tone to be judged.

Returns: whether tone allows a light foreground.